prod fixes

+++ scripts/install.sh
@@ -3,17 +3,19 @@
 set -euo pipefail
 
 # packages
-export DEBIAN_FRONTEND="noninteractive"
-# running this set -u sometimes causes issues with packaging scripts, it seems
-set +u
-sed 's/^#.*//' purge-packages.txt | xargs apt-get purge -y
-apt-get update
-apt-get upgrade -y
-sed 's/^#.*//' core-packages.txt | xargs apt-get install -y
-sed 's/^#.*//' packages.txt | xargs apt-get install -y
-apt-get autoremove -y
-
-set -u
+INSTALL_PACKAGES=${INSTALL_PACKAGES:-true}
+# we can't install packages in EMIS atm, so allow skipping that
+if test "$INSTALL_PACKAGES" = "true"; then
+    # running this set -u sometimes causes issues with packaging scripts, it seems
+    set +u
+    sed 's/^#.*//' purge-packages.txt | xargs apt-get purge -y
+    apt-get update
+    apt-get upgrade -y
+    sed 's/^#.*//' core-packages.txt | xargs apt-get install -y
+    sed 's/^#.*//' packages.txt | xargs apt-get install -y
+    apt-get autoremove -y
+    set -u
+fi
 
 # ensure groups
 for group in developers researchers reviewers; do
@@ -22,7 +24,6 @@ for group in developers researchers reviewers; do
         groupadd $group
     fi
 done
-
 
 # system configurations
